pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "ml_ops_assignment"
version = "0.0.1"
description = "not written yet"
authors = [
  { name = "group_63", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.4.2",
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "fastapi==0.115.6",
    "loguru>=0.7.3",
    "markdown>=3.10",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "torch==2.6.0; sys_platform != 'darwin' or platform_machine == 'arm64'",
    "transformers>=4.57.5",
    "typer==0.12.5",
    "uvicorn==0.34.0",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mypy==1.19.1",
    "mkdocs-material==9.7.1",
    "mkdocstrings-python==2.0.1",
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "dvc-gs>=3.0.2",
    "pytest-cov>=6.3.0",
    "types-pyyaml>=6.0.12.20250915",
]


[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v"

[tool.mypy]
files = ["src", "tests", "tasks.py"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
